chore: move type file to @data folder
