+++ src/modules/feature/room/RoomParticipants/AddParticipants.vue
@@ -76,7 +76,7 @@
 import { useI18n } from "vue-i18n";
 import { AUTH_MODULE_KEY, injectStrict } from "@/utils/inject";
 import { computed, PropType, ref } from "vue";
-import { Participants } from "./types";
+import { Participants } from "../../../data/room/roomParticipants/types";
 
 defineProps({
 	userList: {
