+++ src/modules/feature/room/RoomParticipants/ParticipantsTable.unit.ts
@@ -4,7 +4,7 @@ import {
 } from "@@/tests/test-utils/setup";
 import ParticipantsTable from "./ParticipantsTable.vue";
 import { participants } from "../../../data/room/roomParticipants/mockParticipantsList";
-import { Participants } from "./types";
+import { Participants } from "../../../data/room/roomParticipants/types";
 import { Ref } from "vue";
 import { mdiMenuDown, mdiMenuUp, mdiMagnify } from "@icons/material";
